+++ src/main/java/de/samply/exporter/response/entity/RequestResponseEntity.java
@@ -1,4 +1,4 @@
-package de.samply.teiler.response.entity;
+package de.samply.exporter.response.entity;
 
 public record RequestResponseEntity(
     String responseUrl